+++ lib/deployment_status_slack_notifier.rb
@@ -23,12 +23,15 @@ post '/handler' do
     success_webhooks = ENV['SUCCESS_WEBHOOKS'].split(',')
     failure_webhooks = ENV['FAILURE_WEBHOOKS'].split(',')
     error_webhooks = ENV['ERROR_WEBHOOKS'].split(',')
+    skip_label_name = ENV['SKIP_LABEL'] if ENV['SKIP_LABEL']
+
+    skip_success_notification = skip_label?(payload, skip_label_name) if skip_label_name
 
     slack = DeploymentStatusSlackNotifier::SlackClient.new(success_webhooks, failure_webhooks, error_webhooks)
     case payload['deployment_status']['state']
     when 'success'
       msg = success_message_body(payload)
-      slack.success_message(msg)
+      slack.success_message(msg) unless skip_success_notification
     when 'failure'
       msg = success_message_body(payload)
       slack.failure_message(msg)
@@ -68,4 +71,13 @@ def error_message_body(payload)
   repo_name = payload['repository']['name']
   version = payload['deployment']['ref']
   "UNKNOWN ERROR on Release: #{repo_name} version #{version}"
+end
+
+def skip_label?(_payload, label)
+  return false unless label
+
+  label_count = pull_request['labels'].select { |l| l['name'] =~ /^#{label}/i }
+  return true if label_count == 1
+
+  false
 end