+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This file is used to list changes made in Deployment_status Slack Notifier.
 
 ## Unreleased
 
+- add a `SKIP_LABEL` env var to skip success notifications on a specific label
+
 ## 1.1.1 - *2020-11-18*
 
 - Fixed bug with dockerhub push due to set-env deprecation